change routing for changelist page
